server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
transformers==4.44.2
torch>=2.1.0
Pillow==10.3.0
requests==2.32.3
python-multipart==0.0.9
